requirements.txt
asgiref==3.4.1
crispy-bootstrap5==0.6
Django==3.2.8
django-allauth==0.41.0
django-countries==7.2.1
django-crispy-forms==1.13.0
django-tinymce==3.3.0
oauthlib==3.1.1
Pillow==8.4.0
python3-openid==3.2.0
pytz==2021.3
requests-oauthlib==1.3.0
sqlparse==0.4.2
stripe==2.61.0
